TradeArt.CaseStudy.Business/Implementations/CaseStudyBusiness.cs
using TradeArt.CaseStudy.Business.Interfaces;
using TradeArt.CaseStudy.Model;
using TradeArt.CaseStudy.Model.Requests.CaseStudy;

namespace TradeArt.CaseStudy.Business.Implementations;

public class CaseStudyBusiness : ICaseStudyBusiness {
	public BaseResult InvertText(InvertRequest request) {
		if (string.IsNullOrWhiteSpace(request.Text))
			return new ErrorResult("The Text field is required.");

		var charArray = request.Text.ToCharArray();
		Array.Reverse(charArray);
		return new SuccessResult<string>(new string(charArray));
	}
}